feat: SearchBar
