--- src/Components/Filter/index.jsx
@@ -1,5 +1,0 @@
-const Filter = () => {
-    return null;
-};
-
-export default Filter;
