+++ src/Components/Table/index.jsx
@@ -2,45 +2,7 @@ import { useMemo, useState, useEffect } from 'react';
 import { useTable } from 'react-table';
 import makeData from './data';
 import '@Styles/table.css';
-
-const dummyColumns = () => [
-    {
-        Header: 'Col1',
-        accessor: 'Col1',
-    },
-    {
-        Header: 'Col2',
-        accessor: 'Col2',
-    },
-
-    {
-        Header: 'Col3',
-        accessor: 'Col3',
-    },
-    {
-        Header: 'Col4',
-        accessor: 'Col4',
-    },
-];
-
-const dummyData = () => [
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
-];
+import { dummyColumns, dummyData } from '@Dummy/tableDummy';
 
 const Table = props => {
     // TODO_P :: namor => 이거 패키지 삭제해야함 (랜덤명 생성 패키지)
