+++ src/Dummy/tableDummy.js
@@ -1,0 +1,40 @@
+const dummyColumns = () => [
+    {
+        Header: 'Col1',
+        accessor: 'Col1',
+    },
+    {
+        Header: 'Col2',
+        accessor: 'Col2',
+    },
+
+    {
+        Header: 'Col3',
+        accessor: 'Col3',
+    },
+    {
+        Header: 'Col4',
+        accessor: 'Col4',
+    },
+];
+
+const dummyData = () => [
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+    { Col1: '안녕', Col2: '안녕', Col3: '안녕', Col4: '안녕' },
+];
+
+export { dummyColumns, dummyData };
